#120 Removed install and environment with generic solution
