+++ bin/common/executor/executions/powershell-command/index.js
@@ -1,5 +1,5 @@
 import powershell from '../../../helper/powershell.js';
-import customOptions from '../../../helper/custom_options.js';
+import customOptions from '../../../helper/options/custom-options-creator.js';
 
 import {Execute, Runtime} from "../../../models/dever-json/internal.js";
 import {ExecutionInterface, Result} from "../../models.js";
@@ -21,7 +21,7 @@ export default new class extends ExecutionInterface {
      */
     async handle(execute, runtime) {
         try {
-            const command = customOptions.addOptionsToCommand(execute.command, execute.options, runtime.args);
+            const command = customOptions.addToCommand(execute.command, execute.options, runtime.args);
             await powershell.executeSync(command, execute.runAsElevated);
 
             return this._success(Operation.Executed);
